# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /course-management/students:
    get:
      operationId: listStudents
      tags: [course-management]
      summary: List all registered students
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [course-management]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /course-management/students/{studentId}:
    get:
      operationId: getStudent
      tags: [course-management]
      summary: Get a registered student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [course-management]
      summary: Update a registered student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found

  # ── Audit trail (view-duration reporting) ───────────────────────────────
...
components:
  schemas:
...
    EmployeeStatus:
      type: string
      enum: [active, inactive, on_leave, leaver]
...
    Department:
      type: object
      required: [id, name, description, headEmployeeId, employeeCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        headEmployeeId:
          type: ["integer", "null"]
        employeeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        headEmployeeId:
          type: ["integer", "null"]
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        headEmployeeId:
          type: ["integer", "null"]
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1

    # ── SysAdmin schemas ────────────────────────────────────────────────────
    Permission:
      type: string
      enum:
        - sysadmin
        - hr:access
        - hr:past_employees
        - hr_admin
        - view_employees
        - edit_employees
        - delete_employees
        - view_departments
        - edit_departments
        - view_leave
        - manage_leave
        - view_reports
        - view_payroll
        - view_disclosures
        - review_disclosures
        - view_own_profile
        - upload_qualifications
        - view_employee_directory
    UserStatus:
      type: string
      description: User account status value — managed via List of Values
...
    UserLinkedEmployee:
      type: object
      required: [id, firstName, lastName, status]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        status:
          $ref: "#/components/schemas/EmployeeStatus"
    RoleSummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    User:
      type: object
      required: [id, name, email, status, roles, permissions, isSystemAccount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        status:
          $ref: "#/components/schemas/UserStatus"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleSummary"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        isSystemAccount:
          type: boolean
        employeeId:
          type: ["integer", "null"]
        employee:
          oneOf:
            - $ref: "#/components/schemas/UserLinkedEmployee"
            - type: "null"
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Student:
      type: object
      required: [studentId, firstName, lastName]
      properties:
        studentId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        homeAddress:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        emailAddress:
          type: ["string", "null"]
    StudentInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        homeAddress:
          type: string
        phoneNumber:
          type: string
        emailAddress:
          type: string
          description: Must be a valid email address — validated server-side.
    StudentUpdate:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        homeAddress:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        emailAddress:
          type: ["string", "null"]
          description: Must be a valid email address — validated server-side.
...